Prevent status_message() from dying if the GUI hasn't been fully set up yet. Fixes http://bugs.activestate.com/show_bug.cgi?id=47678
[git-p4: depot-paths = "//depot/main/Camel/src/as-mod/ActivePerl-PPM/main/": change = 265827]

+++ lib/ActivePerl/PPM/GUI.pm
@@ -1247,13 +1247,15 @@ sub status_message {
     my $ins = delete $opts{insert} || "end";
     my $clr = delete $opts{clear} || 0;
 
-    $pw_nb->select($status_sw);
-    $status_box->configure(-state => "normal");
-    $status_box->delete("1.0", "end") if $clr;
-    $status_box->insert($ins, $msg, $tag);
-    $status_box->configure(-state => "disabled");
-    $status_box->see($ins);
-    Tkx::update('idletasks');
+    if (defined($pw_nb) && defined($status_box)) {
+	$pw_nb->select($status_sw);
+	$status_box->configure(-state => "normal");
+	$status_box->delete("1.0", "end") if $clr;
+	$status_box->insert($ins, $msg, $tag);
+	$status_box->configure(-state => "disabled");
+	$status_box->see($ins);
+	Tkx::update('idletasks');
+    }
     if ($ENV{'ACTIVEPERL_PPM_DEBUG'}) {
 	print $msg;
     }
